api/requirements.txt
Flask==1.1.4
Flask-RESTful==0.3.9
Flask-JWT-Extended==3.25.1
flask-swagger-ui==3.36.0
flask-marshmallow==0.14.0
marshmallow==2.21.0
requests==2.27.1
python-dotenv==0.18.0
Werkzeug==1.0.1
MarkupSafe==1.1.1
pytest==4.6.11
pytest-flask==0.15.1
itsdangerous==1.1.0
Jinja2==2.11.3
mock==3.0.5